refactored func
